refactor(terminal): move the pane's pure geometry helpers out of the component
React Doctor flagged them on this PR: a component file that also exports plain
functions makes every consumer of one of them import the whole terminal, and
they are already unit-tested on their own. They move to ghosttyGeometry.ts
unchanged, with their tests renamed to follow.

The construction effect's early return now hands back an explicit no-op
cleanup, so every path out of it returns one rather than leaving the reader
(and the analyzer) to prove nothing was allocated yet.
